- Moved CLI to separate file - Fixed dependencies (?)
