+++ Makefile
@@ -27,7 +27,7 @@ requirements: test_environment
 
 ## Make Dataset
 data: requirements
-	$(PYTHON_INTERPRETER) src/data/make_dataset.py data/raw data/processed
+	$(PYTHON_INTERPRETER) src/cli/main.py data/raw data/processed
 
 ## Delete all compiled Python files
 clean:
