+++ requirements.txt
@@ -9,4 +9,6 @@ awscli
 flake8
 python-dotenv>=0.5.1
 wget==3.2
-unzip==0.1
+unzip==1.0.0
+
+setuptools~=67.7.2
